Use a different endpoint URL for reading objects.
This is the URL that is described in the "Request Endpoints" section of
the "Concepts and Techniques" document, rather than the one used on the
Objects.get page that we were previously using.

I happen to know from personal communication (and from verifying with
Google-internal headers enabled) that it takes a more direct route to
the backend. Hopefully it is lower latency for this reason.