Assign new generation numbers when copying in gcsfake.
This upholds the real GCS guarantee that the generation number for a
particular name cannot regress. This was coming up in a gcsfuse
integration that did the following:

1.  Create 'foo'.
2.  Create 'bar'.
3.  Rename 'foo' on top of 'bar'.

This resulted in the generation number for 'bar' going backwards, which
violated a file system assumption and caused the inode lookup loop to
not converge.

For GoogleCloudPlatform/gcsfuse#81.