Fixed a range read bug where HTTP connections would not be reused.
The implementation would previously stop reading from the HTTP response
body when it hit the end of the range of interest. Since we ask GCS to
overread by a byte in order to work around HTTP protocol limitations,
this would cause us to always abandon the body before it saw EOF,
preventing the http package from reusing the connection.

Now we read the extra byte and throw it away.