Add flag ReadCompressed to gcs.ReadObjectRequest (#34)
This is a boolean flag added to gcs.ReadObjectRequest to allow caller to
indicate the need to read the object content as compressed from the
bucket interface. It is meant only for objects which have
'content-encoding' set to 'gzip' in their GCS metadata.
The bucket implementation can accordingly take care of creating the
read-request or of handling the request-response on the basis of this
flag.